extract "focus input" functionality into a mixin
