+++ core/client/tests/unit/components/gh-trim-focus-input-test.js
